// src/app/common/model/main.ts
export class ExceptionList {
  time: string;
  type: string;
  position: string;
  description: string;
}
export class PageData {
  pipeCircuitId?: any;
  state?: any;
  currentPage?: any;
  pageSize?: any;
}
export class HisData {
  pipeCircuitId?: any;
  timeMarking?: any;
  currentPage?: any;
  pageSize?: any;
}
export class TBData {
  pipeCircuitId?: any;
  currentPage?: any;
  pageSize?: any;
}
export class RepairData {
  circuitId?: any;
}